requirements.txt
httpx>=0.27.0
flask>=3.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
python-telegram-bot>=21.0
anthropic>=0.34.0
schedule>=1.2.0
python-dotenv>=1.0.0
lxml>=5.2.0
